app/requirements.txt
Flask
pandas
numpy
scikit-surprise
streamlit
jupyterlab